+++ src/east/modules/tsuite.py
@@ -1,5 +1,5 @@
 import os
-from typing import NamedTuple, Sequence
+from typing import NamedTuple
 
 from ..constants import EAST_GITHUB_URL